CLI changes and test fixes
 - Make the URL argument optional and default to peer-to-peer if it's absent
 - Fix duration handling
 - Update brokerlib

+++ python/brokerlib.py
@@ -1,3 +1,4 @@
+#
 # Licensed to the Apache Software Foundation (ASF) under one
 # or more contributor license agreements.  See the NOTICE file
 # distributed with this work for additional information
@@ -16,12 +17,6 @@
 # under the License.
 #
 
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-from __future__ import unicode_literals
-from __future__ import with_statement
-
 import collections as _collections
 import os as _os
 import proton as _proton
@@ -33,12 +28,12 @@ import subprocess as _subprocess
 import sys as _sys
 import time as _time
 import tempfile as _tempfile
-import pathlib as _pathlib
 
-class Broker(object):
-    def __init__(self, scheme, host, port, id=None, user=None, password=None, ready_file=None,
-                 cert=None, key=None, key_password=None, trusted_db=None):
-        self.scheme = scheme
+class Broker:
+    def __init__(self, host, port, id=None, ready_file=None,
+                 user=None, password=None,
+                 cert=None, key=None, trust_store=None,
+                 quiet=False, verbose=False, init_only=False):
         self.host = host
         self.port = port
         self.id = id
@@ -47,35 +42,44 @@ class Broker(object):
         self.ready_file = ready_file
         self.cert = cert
         self.key = key
-        self.key_password = key_password
-        self.trusted_db = trusted_db
+        self.trust_store = trust_store
+        self.quiet = quiet
+        self.verbose = verbose
+        self.init_only = init_only
 
         if self.id is None:
-            self.id = "broker-{0}".format(_uuid.uuid4())
+            self.id = "broker-{0}".format(str(_uuid.uuid4())[:8])
 
-        self.container = _reactor.Container(_Handler(self), self.id)
+        self.container = _reactor.Container(_Handler(self))
+        self.container.container_id = self.id # XXX Obnoxious
 
         self._config_dir = None
 
     def init(self):
+        self.info("Initializing {0}", self)
+
         if self.user is not None:
             if self.password is None:
                 self.fail("A password is required for user authentication")
 
             self._init_sasl_config()
 
-        if self.scheme == "amqps":
-            if self.key is None or self.cert is None:
-                self.fail("if scheme is amqps, key and cert files must be specified")
-            if not _pathlib.Path(self.key).is_file():
-                self.fail("key file %s does not exist" % (self.key))
-            if not _pathlib.Path(self.cert).is_file():
-                self.fail("cert file %s does not exist" % (self.cert))
-            if self.trusted_db and not _pathlib.Path(self.trusted_db).is_file():
-                self.fail("trusted db file %s does not exist" % (self.trusted_db))
+        if self.cert is not None:
+            if self.key is None:
+                self.fail("Both the cert and key files must be provided")
+
+            if not _os.path.isfile(self.key):
+                self.fail("Private key file {0} does not exist", self.key)
+
+            if not _os.path.isfile(self.cert):
+                self.fail("Certificate file {0} does not exist", self.cert)
+
+            if self.trust_store and not _os.path.isfile(self.trust_store):
+                self.fail("Trust store file {0} does not exist", self.trust_store)
 
     def _init_sasl_config(self):
         self._config_dir = _tempfile.mkdtemp(prefix="brokerlib-", suffix="")
+
         config_file = _os.path.join(self._config_dir, "proton-server.conf")
         sasldb_file = _os.path.join(self._config_dir, "users.sasldb")
 
@@ -93,6 +97,9 @@ class Broker(object):
         except _subprocess.CalledProcessError as e:
             self.fail("Failed adding user to SASL database: {0}", e)
 
+    def debug(self, message, *args):
+        pass
+
     def info(self, message, *args):
         pass
 
@@ -103,20 +110,31 @@ class Broker(object):
         pass
 
     def error(self, message, *args):
-        _sys.stderr.write("{0}\n".format(message.format(*args)))
-        _sys.stderr.flush()
+        self.log(message, *args)
 
     def fail(self, message, *args):
         self.error(message, *args)
         _sys.exit(1)
 
+    def log(self, message, *args):
+        message = message[0].upper() + message[1:]
+        message = message.format(*args)
+        message = "{0}: {1}".format(self.id, message)
+
+        _sys.stderr.write("{0}\n".format(message))
+        _sys.stderr.flush()
+
     def run(self):
-        self.container.run()
+        try:
+            if self.init_only:
+                return
 
-        if _os.path.exists(self._config_dir):
-            _shutil.rmtree(self.dir, ignore_errors=True)
+            self.container.run()
+        finally:
+            if self._config_dir and _os.path.exists(self._config_dir):
+                _shutil.rmtree(self.dir, ignore_errors=True)
 
-class _Queue(object):
+class _Queue:
     def __init__(self, broker, address):
         self.broker = broker
         self.address = address
@@ -135,7 +153,7 @@ class _Queue(object):
 
         self.consumers.append(link)
 
-        self.broker.info("Added consumer for {0} to {1}", link.connection, self)
+        self.broker.info("Added consumer for {0} to {1}", _container_repr(link.connection), self)
 
     def remove_consumer(self, link):
         assert link.is_sender
@@ -145,12 +163,12 @@ class _Queue(object):
         except ValueError:
             return
 
-        self.broker.info("Removed consumer for {0} from {1}", link.connection, self)
+        self.broker.info("Removed consumer for {0} from {1}", _container_repr(link.connection), self)
 
     def store_message(self, delivery, message):
         self.messages.append(message)
 
-        self.broker.notice("Stored {0} from {1} on {2}", message, delivery.connection, self)
+        self.broker.notice("Stored {0} from {1} on {2}", message, _container_repr(delivery.connection), self)
 
     def forward_messages(self):
         credit = sum([x.credit for x in self.consumers])
@@ -173,7 +191,7 @@ class _Queue(object):
                 consumer.send(message)
                 sent += 1
 
-                self.broker.notice("Forwarded {0} on {1} to {2}", message, self, consumer.connection)
+                self.broker.notice("Forwarded {0} on {1} to {2}", message, self, _container_repr(consumer.connection))
 
         self.consumers.rotate(sent)
 
@@ -186,14 +204,16 @@ class _Handler(_handlers.MessagingHandler):
         self.verbose = False
 
     def on_start(self, event):
-        interface = "{0}://{1}:{2}".format(self.broker.scheme, self.broker.host, self.broker.port)
+        interface = "{0}:{1}".format(self.broker.host, self.broker.port)
 
-        if self.broker.scheme == "amqps":
+        if self.broker.cert is not None:
             server_ssl_domain = event.container.ssl.server
-            server_ssl_domain.set_credentials(self.broker.cert, self.broker.key, self.broker.key_password)
-            if self.broker.trusted_db:
-                server_ssl_domain.set_trusted_ca_db(self.broker.trusted_db)
-                server_ssl_domain.set_peer_authentication(_proton.SSLDomain.VERIFY_PEER, self.broker.trusted_db)
+            server_ssl_domain.set_credentials(self.broker.cert, self.broker.key, None)
+
+            if self.broker.trust_store:
+                server_ssl_domain.set_trusted_ca_db(self.broker.trust_store)
+                server_ssl_domain.set_peer_authentication(_proton.SSLDomain.VERIFY_PEER_NAME,
+                                                          self.broker.trust_store)
             else:
                 server_ssl_domain.set_peer_authentication(_proton.SSLDomain.ANONYMOUS_PEER)
 
@@ -202,7 +222,6 @@ class _Handler(_handlers.MessagingHandler):
         self.broker.notice("Listening for connections on '{0}'", interface)
 
         if self.broker.ready_file is not None:
-            _time.sleep(0.1) # XXX
             with open(self.broker.ready_file, "w") as f:
                 f.write("ready\n")
 
@@ -210,26 +229,52 @@ class _Handler(_handlers.MessagingHandler):
         try:
             queue = self.queues[address]
         except KeyError:
-            queue = self.queues[address] = _Queue(self.broker, address)
+            queue = self.create_queue(address)
 
+        return queue
+
+    def create_queue(self, address):
+        assert address not in self.queues, address
+
+        queue = _Queue(self.broker, address)
+        self.queues[address] = queue
         return queue
 
     def on_link_opening(self, event):
         if event.link.is_sender:
+            # A client receiving from the broker
+
             if event.link.remote_source.dynamic:
+                # A temporary queue
                 address = "{0}/{1}".format(event.connection.remote_container, event.link.name)
+                queue = self.create_queue(address)
+            elif event.link.remote_source.address in (None, ""):
+                raise Exception("The client created a receiver with no source address")
             else:
+                # A named queue
                 address = event.link.remote_source.address
+                queue = self.get_queue(address)
 
             assert address is not None
 
             event.link.source.address = address
-
-            queue = self.get_queue(address)
             queue.add_consumer(event.link)
 
         if event.link.is_receiver:
-            address = event.link.remote_target.address
+            # A client sending to the broker
+
+            if event.link.remote_target.dynamic:
+                # A temporary queue
+                address = "{0}/{1}".format(event.connection.remote_container, event.link.name)
+                queue = self.create_queue(address)
+            elif event.link.remote_target.address in (None, ""):
+                # Anonymous relay - no queueing
+                address = None
+            else:
+                # A named queue
+                address = event.link.remote_target.address
+                queue = self.get_queue(address)
+
             event.link.target.address = address
 
     def on_link_closing(self, event):
@@ -237,24 +282,21 @@ class _Handler(_handlers.MessagingHandler):
             queue = self.queues[event.link.source.address]
             queue.remove_consumer(event.link)
 
-    def on_connection_init(self, event):
-        event.transport.sasl().allow_insecure_mechs=True
-
     def on_connection_opening(self, event):
         # XXX I think this should happen automatically
         event.connection.container = event.container.container_id
 
     def on_connection_opened(self, event):
-        self.broker.notice("Opened connection from {0}", event.connection)
+        self.broker.notice("Opened connection from {0}", _container_repr(event.connection))
 
     def on_connection_closing(self, event):
         self.remove_consumers(event.connection)
 
     def on_connection_closed(self, event):
-        self.broker.notice("Closed connection from {0}", event.connection)
+        self.broker.notice("Closed connection from {0}", _container_repr(event.connection))
 
     def on_disconnected(self, event):
-        self.broker.notice("Disconnected from {0}", event.connection)
+        self.broker.notice("Disconnected from {0}", _container_repr(event.connection))
 
         self.remove_consumers(event.connection)
 
@@ -277,62 +319,126 @@ class _Handler(_handlers.MessagingHandler):
         queue.forward_messages()
 
     def on_settled(self, event):
-        template = "Container '{0}' {1} {2} to {3}"
+        template = "Container '{0}' {1} {2} for {3}"
         container = event.connection.remote_container
-        source = event.link.source
+        source = _terminus_repr(event.link.source)
         delivery = event.delivery
 
         if delivery.remote_state == delivery.ACCEPTED:
-            self.broker.info(template, container, "accepted", delivery, source)
+            self.broker.info(template, container, "accepted", _delivery_repr(delivery), source)
         elif delivery.remote_state == delivery.REJECTED:
-            self.broker.warn(template, container, "rejected", delivery, source)
+            self.broker.warn(template, container, "rejected", _delivery_repr(delivery), source)
         elif delivery.remote_state == delivery.RELEASED:
-            self.broker.notice(template, container, "released", delivery, source)
+            self.broker.notice(template, container, "released", _delivery_repr(delivery), source)
         elif delivery.remote_state == delivery.MODIFIED:
-            self.broker.notice(template, container, "modified", delivery, source)
+            self.broker.notice(template, container, "modified", _delivery_repr(delivery), source)
 
     def on_message(self, event):
         message = event.message
         delivery = event.delivery
         address = event.link.target.address
 
-        if address is None:
+        if address in (None, ""):
             address = message.address
 
         queue = self.get_queue(address)
         queue.store_message(delivery, message)
         queue.forward_messages()
-    #
-    # def on_unhandled(self, name, event):
-    #     _sys.stderr.write("{0} {1}\n".format(name, event))
-    #     _sys.stderr.flush()
 
-if __name__ == "__main__":
-    def _print(message, *args):
-        message = message.format(*args)
-        _sys.stderr.write("{0}\n".format(message))
-        _sys.stderr.flush()
+    def on_unhandled(self, name, event):
+        self.broker.debug("Unhandled event: {0} {1}", name, event)
+
+def _container_repr(connection):
+    return "container '{0}'".format(connection.remote_container)
+
+def _terminus_repr(terminus):
+    return "terminus '{0}'".format(terminus.address)
+
+def _delivery_repr(delivery):
+    return "delivery '{0}'".format(delivery.tag)
+
+def wait_for_broker(ready_file, timeout=30):
+    start_time = _time.time()
+    interval = 0.125
+
+    while True:
+        if _time.time() - start_time > timeout:
+            raise Exception("Timed out waiting for the broker")
+
+        _time.sleep(interval)
+
+        with open(ready_file, "r") as f:
+            if f.read() == "ready\n":
+                break
+
+        if interval < 1:
+            interval = interval * 2
+        else:
+            print("Still waiting for the broker")
+
+def main():
+    import argparse
+
+    parser = argparse.ArgumentParser(description="An AMQP message broker for testing")
+
+    parser.add_argument("--host", metavar="HOST", default="localhost",
+                        help="Listen for connections on HOST (default localhost)")
+    parser.add_argument("--port", metavar="PORT", default=5672, type=int,
+                        help="Listen for connections on PORT (default 5672)")
+    parser.add_argument("--id", metavar="ID",
+                        help="Set the container identity to ID (default is generated)")
+    parser.add_argument("--ready-file", metavar="FILE",
+                        help="The file used to indicate the server is ready")
+    # parser.add_argument("--user", metavar="USER",
+    #                   help="Require USER")
+    # parser.add_argument("--password", metavar="SECRET",
+    #                   help="Require SECRET")
+    parser.add_argument("--cert", metavar="FILE",
+                        help="The TLS certificate file")
+    parser.add_argument("--key", metavar="FILE",
+                        help="The TLS private key file")
+    parser.add_argument("--trust-store", metavar="FILE",
+                        help="The file containing trusted client certificates.  "
+                        "If set, the server verifies client identities.")
+    parser.add_argument("--quiet", action="store_true",
+                        help="Print no logging to the console")
+    parser.add_argument("--verbose", action="store_true",
+                        help="Print detailed logging to the console")
+    parser.add_argument("--init-only", action="store_true",
+                      help=argparse.SUPPRESS)
+
+    args = parser.parse_args()
 
     class _Broker(Broker):
-        def info(self, message, *args): _print(message, *args)
-        def notice(self, message, *args): _print(message, *args)
-        def warn(self, message, *args): _print(message, *args)
+        def debug(self, message, *args):
+            if self.verbose:
+                self.log(message, *args)
 
-    try:
-        host, port = _sys.argv[1:3]
-    except IndexError:
-        _print("Usage: brokerlib <host> <port>")
-        _sys.exit(1)
+        def info(self, message, *args):
+            if self.verbose:
+                self.log(message, *args)
 
-    try:
-        port = int(port)
-    except ValueError:
-        _print("The port must be an integer")
-        _sys.exit(1)
+        def notice(self, message, *args):
+            if not self.quiet:
+                self.log(message, *args)
 
-    broker = _Broker(host, port)
+        def warn(self, message, *args):
+            message = "Warning! {0}".format(message)
+            self.log(message, *args)
+
+        def error(self, message, *args):
+            message = "Error! {0}".format(message)
+            self.log(message, *args)
+
+    broker = _Broker(args.host, args.port, id=args.id, ready_file=args.ready_file,
+                     # user=args.user, password=args.password,
+                     cert=args.cert, key=args.key, trust_store=args.trust_store,
+                     quiet=args.quiet, verbose=args.verbose, init_only=args.init_only)
 
     try:
         broker.run()
     except KeyboardInterrupt:
         pass
+
+if __name__ == "__main__":
+    main()
